[FEAT] Add basic structure for state inheritance
